De-couple schema loading from Schema generation
This could be overkill but I thought it would be better to de-couple
schema loading from schema generation. One future idea is to allow
schema loading from multiple serialization formats. Also I thought
this is a good use-case to de-couple I/O from business logic. So it
should not matter where the schema is coming from